05-asmiov-audio-transcripts/requirements.txt
ipykernel
openai
python-dotenv
streamlit
streamlit_webrtc
moviepy
pydub
